+++ front/app/api/graph_data_units/responseTypes/VisitorsTrafficSourcesWidget.ts
@@ -1,0 +1,19 @@
+export interface TrafficSourcesRow {
+  count: number;
+  'dimension_referrer_type.id': string;
+  first_dimension_referrer_type_name: ReferrerTypeName;
+}
+
+export type ReferrerTypeName =
+  | 'Direct Entry'
+  | 'Social Networks'
+  | 'Search Engines'
+  | 'Websites'
+  | 'Campaigns';
+
+export interface VisitorsTrafficSourcesResponse {
+  data: {
+    type: 'report_builder_data_units';
+    attributes: TrafficSourcesRow[];
+  };
+}
