+++ front/app/components/admin/GraphCards/VisitorsTrafficSourcesCard/useVisitorReferrerTypes/typings.ts
@@ -1,16 +1,3 @@
-export interface TrafficSourcesRow {
-  count: number;
-  'dimension_referrer_type.id': string;
-  first_dimension_referrer_type_name: ReferrerTypeName;
-}
-
-export type ReferrerTypeName =
-  | 'Direct Entry'
-  | 'Social Networks'
-  | 'Search Engines'
-  | 'Websites'
-  | 'Campaigns';
-
 // Hook return value
 export interface PieRow {
   name: string;
